Change calculation number attributes to numbers
The previousNumber, currentNumber, and result are stored as strings
because they are taken from the textContent of the input field.

It makes more sense to store them as numbers by converting the
textContent to a number when initially storing the currentNumber so that
calculation functions can be done on the native variables without
converting to numbers each time.