Add clearCalculation function to clear button
The clear button only clears the input field but the calculation object
still contains the information from the previous calculation.

This leads to potential errors such as pressing the equal sign with
one or no value in the input field.

These errors can be prevented by resetting all calculation values.