Change when input field is cleared
Instead of having the input field cleared when an operator button is
clicked, pressing a number button will clear the input field if
isWaitingForNumber is true.

This attribute will be set to true when the equal button is pressed as
well, so that the screen will clear if a result is being displayed and
a number button is pressed to start a new calculation.