Allow calculation to be repeated with equal button
Use a new calculation attribute called repeatNumber which stores the
value of the number used in the previous operation. This allows the
user to repeatedly click the equal button repeat the same calculation
on the new result. This is reset when the calculation is cleared.

For example, if the user calculates some x * 2 and presses the equal
button again, then the result should be x * 2 * 2 and so on until new
numbers are introduced or the calculator is cleared.